requirements.txt
git+https://github.com/Rapptz/discord.py
discord
asyncio
sys
pillow
ordinal
io